Will refactor utils
- split utils methods be common, date, render files
- the render function can accept components
