+++ src/main.js
@@ -11,8 +11,9 @@ import Comments from './view/comments.js';
 import NewCommentView from './view/new-comment.js';
 import {generateFilm} from './mock/mock-film.js';
 import {generateComment} from './mock/mock-comment.js';
-import {render, sortFilmByComments, sortFilmByRating} from './utils';
+import {sortFilmByComments, sortFilmByRating} from './utils/common.js';
 import {NavigationTypes, FilmListTypes} from './utils/const.js';
+import {render} from './utils/render.js';
 
 const CARDS_LOAD_STEP = 5;
 const CARDS_EXTRA_LOAD_STEP = 2;
@@ -67,10 +68,10 @@ const renderCards = (container, films) => {
       const popupElement = popupComponent.getElement();
       const filmDetailBottomContainer = popupElement.querySelector('.film-details__bottom-container');
 
-      render(filmDetailBottomContainer, new Comments(getFilmComments(film)).getElement());
+      render(filmDetailBottomContainer, new Comments(getFilmComments(film)));
 
       const commentsContainer = filmDetailBottomContainer.querySelector('.film-details__comments-wrap');
-      render(commentsContainer, new NewCommentView().getElement());
+      render(commentsContainer, new NewCommentView());
 
       const closeButtonElement = popupElement.querySelector('.film-details__close-btn');
       closeButtonElement.addEventListener('click', onPopupCloseHandler, {once: true});
@@ -83,16 +84,16 @@ const renderCards = (container, films) => {
   });
 };
 
-render(headerElement, new UserProfileView(watchedFilmsCount).getElement());
-render(mainElement, new MainNavigationView(navigationStatistics).getElement());
-render(mainElement, new SortMenuView().getElement());
-render(mainElement, new FilmsBoardView().getElement());
+render(headerElement, new UserProfileView(watchedFilmsCount));
+render(mainElement, new MainNavigationView(navigationStatistics));
+render(mainElement, new SortMenuView());
+render(mainElement, new FilmsBoardView());
 
 const filmBoard = document.querySelector('.films');
 const filmListAll = new FilmListView(FilmListTypes.ALL_MOVIES);
-render(filmBoard, filmListAll.getElement());
-render(filmBoard, new FilmListView(FilmListTypes.TOP_MOVIES).getElement());
-render(filmBoard, new FilmListView(FilmListTypes.COMMENTED_MOVIES).getElement());
+render(filmBoard, filmListAll);
+render(filmBoard, new FilmListView(FilmListTypes.TOP_MOVIES));
+render(filmBoard, new FilmListView(FilmListTypes.COMMENTED_MOVIES));
 
 const allMovieSection = filmBoard.querySelector('.films-list');
 const allMoviesList = allMovieSection.querySelector('.films-list__container');
@@ -101,11 +102,11 @@ renderCards(allMoviesList, filmsData.slice(0, CARDS_LOAD_STEP));
 renderCards(topMoviesList, topFilmsData.slice(0, CARDS_EXTRA_LOAD_STEP));
 renderCards(commentMoviesList, commentsFilmsData.slice(0, CARDS_EXTRA_LOAD_STEP) );
 
-render(footerStatisticsElement, new FooterStatisticsView(filmsData.length).getElement());
+render(footerStatisticsElement, new FooterStatisticsView(filmsData.length));
 
 if (filmsData.length > CARDS_LOAD_STEP) {
   let renderedCardCount = CARDS_LOAD_STEP;
-  render(allMovieSection, new ShowMoreButtonView().getElement());
+  render(allMovieSection, new ShowMoreButtonView());
   const loadMoreButton = allMovieSection.querySelector('.films-list__show-more');
 
   loadMoreButton.addEventListener('click', (evt) => {
