+++ src/utils/const.js
@@ -38,13 +38,7 @@ const SortTypes = {
   RATING: 'rating',
 };
 
-const RenderPosition = {
-  AFTEREND: 'afterend',
-  BEFOREEND: 'beforeend',
-};
-
 export {
-  RenderPosition,
   FilmListTypes,
   emotions,
   FilmControlTypes,
